Migrate to using Next.js 13 and the new app directory
