--- .nvmrc
@@ -1,1 +1,0 @@
-16.3.0
