+++ components/LinkList.tsx
@@ -1,4 +1,4 @@
-import Link from './Link';
+import Link from '@/components/Link';
 
 interface LinkProps {
   href: string;
