+++ components/ThemeProvider.tsx
@@ -1,0 +1,5 @@
+'use client';
+
+import { ThemeProvider } from 'next-themes';
+
+export default ThemeProvider;
