+++ next.config.js
@@ -1,3 +1,8 @@
-module.exports = {
-  reactStrictMode: true,
-};
+/** @type {import('next').NextConfig} */
+const nextConfig = {
+  experimental: {
+    appDir: true,
+  },
+}
+
+module.exports = nextConfig
