+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "jagd.me",
-  "version": "2.0.0",
+  "version": "3.0.0",
   "private": true,
   "repository": {
     "type": "git",
@@ -11,25 +11,23 @@
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "lint": "next lint",
-    "format": "prettier --write ."
+    "lint": "next lint"
   },
   "dependencies": {
-    "next": "^11.1.3",
-    "next-themes": "^0.0.15",
-    "react": "17.0.2",
-    "react-dom": "17.0.2"
+    "@types/node": "18.15.11",
+    "@types/react": "18.0.31",
+    "@types/react-dom": "18.0.11",
+    "eslint": "8.37.0",
+    "eslint-config-next": "13.2.4",
+    "next": "13.2.4",
+    "next-themes": "^0.2.1",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "typescript": "5.0.2"
   },
   "devDependencies": {
-    "@types/react": "17.0.11",
-    "autoprefixer": "^10.4.1",
-    "cssnano": "^5.0.14",
-    "eslint": "7.29.0",
-    "eslint-config-next": "11.0.0",
-    "eslint-config-prettier": "^8.3.0",
-    "postcss": "^8.2.15",
-    "prettier": "^2.3.1",
-    "tailwindcss": "^3.0.8",
-    "typescript": "4.3.4"
+    "autoprefixer": "^10.4.14",
+    "postcss": "^8.4.21",
+    "tailwindcss": "^3.3.0"
   }
 }
